# Adopt WellKnownTypeProvider for type lookup by name