--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/WellKnownTypeNames.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/WellKnownTypeNames.cs
@@ -1,9 +1,10 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.CodeAnalysis.Razor;
+namespace Microsoft.AspNetCore.Razor.Language;
 
-internal static class WellKnownTypeNames
+public static class WellKnownTypeNames
 {
+    public const string MicrosoftAspNetCoreComponentsIComponent = "Microsoft.AspNetCore.Components.IComponent";
     public const string SystemThreadingTasksTask1 = "System.Threading.Tasks.Task`1";
 }