--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/WellKnownTypeNames.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/WellKnownTypeNames.cs
@@ -5,6 +5,14 @@ namespace Microsoft.AspNetCore.Razor.Language;
 
 public static class WellKnownTypeNames
 {
+    public const string MicrosoftAspNetCoreComponentsCascadingTypeParameterAttribute = "Microsoft.AspNetCore.Components.CascadingTypeParameterAttribute";
+    public const string MicrosoftAspNetCoreComponentsComponentBase = "Microsoft.AspNetCore.Components.ComponentBase";
+    public const string MicrosoftAspNetCoreComponentsEditorRequiredAttribute = "Microsoft.AspNetCore.Components.EditorRequiredAttribute";
+    public const string MicrosoftAspNetCoreComponentsEventCallback = "Microsoft.AspNetCore.Components.EventCallback";
+    public const string MicrosoftAspNetCoreComponentsEventCallback1 = "Microsoft.AspNetCore.Components.EventCallback`1";
     public const string MicrosoftAspNetCoreComponentsIComponent = "Microsoft.AspNetCore.Components.IComponent";
+    public const string MicrosoftAspNetCoreComponentsParameterAttribute = "Microsoft.AspNetCore.Components.ParameterAttribute";
+    public const string MicrosoftAspNetCoreComponentsRenderFragment = "Microsoft.AspNetCore.Components.RenderFragment";
+    public const string MicrosoftAspNetCoreComponentsRenderFragment1 = "Microsoft.AspNetCore.Components.RenderFragment`1";
     public const string SystemThreadingTasksTask1 = "System.Threading.Tasks.Task`1";
 }