--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/WellKnownTypeNames.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/WellKnownTypeNames.cs
@@ -1,6 +1,8 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.ComponentModel;
+
 namespace Microsoft.AspNetCore.Razor.Language;
 
 public static class WellKnownTypeNames
@@ -14,5 +16,13 @@ public static class WellKnownTypeNames
     public const string MicrosoftAspNetCoreComponentsParameterAttribute = "Microsoft.AspNetCore.Components.ParameterAttribute";
     public const string MicrosoftAspNetCoreComponentsRenderFragment = "Microsoft.AspNetCore.Components.RenderFragment";
     public const string MicrosoftAspNetCoreComponentsRenderFragment1 = "Microsoft.AspNetCore.Components.RenderFragment`1";
+    public const string MicrosoftAspNetCoreRazorTagHelpersHtmlAttributeNameAttribute = "Microsoft.AspNetCore.Razor.TagHelpers.HtmlAttributeNameAttribute";
+    public const string MicrosoftAspNetCoreRazorTagHelpersHtmlAttributeNotBoundAttribute = "Microsoft.AspNetCore.Razor.TagHelpers.HtmlAttributeNotBoundAttribute";
+    public const string MicrosoftAspNetCoreRazorTagHelpersHtmlTargetElementAttribute = "Microsoft.AspNetCore.Razor.TagHelpers.HtmlTargetElementAttribute";
+    public const string MicrosoftAspNetCoreRazorTagHelpersITagHelper = "Microsoft.AspNetCore.Razor.TagHelpers.ITagHelper";
+    public const string MicrosoftAspNetCoreRazorTagHelpersOutputElementHintAttribute = "Microsoft.AspNetCore.Razor.TagHelpers.OutputElementHintAttribute";
+    public const string MicrosoftAspNetCoreRazorTagHelpersRestrictChildrenAttribute = "Microsoft.AspNetCore.Razor.TagHelpers.RestrictChildrenAttribute";
+    public const string SystemCollectionsGenericIDictionary2 = "System.Collections.Generic.IDictionary`2";
+    public const string SystemComponentModelEditorBrowsableAttribute = $"{nameof(System)}.{nameof(System.ComponentModel)}.{nameof(EditorBrowsableAttribute)}";
     public const string SystemThreadingTasksTask1 = "System.Threading.Tasks.Task`1";
 }