--- a/src/Compiler/Microsoft.CodeAnalysis.Razor/src/ComponentDetectionConventions.cs
+++ b/src/Compiler/Microsoft.CodeAnalysis.Razor/src/ComponentDetectionConventions.cs
@@ -1,10 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-#nullable enable
-
 using System;
-using System.Linq;
 
 namespace Microsoft.CodeAnalysis.Razor;
 
@@ -25,24 +22,6 @@ internal static class ComponentDetectionConventions
         return
             symbol.DeclaredAccessibility == Accessibility.Public &&
             !symbol.IsAbstract &&
-            symbol.AllInterfaces.Contains(icomponentSymbol);
-    }
-
-    public static bool IsComponent(INamedTypeSymbol symbol, string icomponentSymbolName)
-    {
-        if (symbol is null)
-        {
-            throw new ArgumentNullException(nameof(symbol));
-        }
-
-        if (icomponentSymbolName is null)
-        {
-            throw new ArgumentNullException(nameof(icomponentSymbolName));
-        }
-
-        return
-            symbol.DeclaredAccessibility == Accessibility.Public &&
-            !symbol.IsAbstract &&
-            symbol.AllInterfaces.Any(s => s.HasFullName(icomponentSymbolName));
+            symbol.AllInterfaces.Contains(icomponentSymbol, SymbolEqualityComparer.Default);
     }
 }