--- a/src/Compiler/Microsoft.CodeAnalysis.Razor/src/ComponentDetectionConventions.cs
+++ b/src/Compiler/Microsoft.CodeAnalysis.Razor/src/ComponentDetectionConventions.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
+using System.Collections.Immutable;
 
 namespace Microsoft.CodeAnalysis.Razor;