… src/Compiler/Microsoft.AspNetCore.Mvc.Razor.Extensions.Version1_X/test/ViewComponentTagHelperDescriptorProviderTest.cs
… src/Compiler/Microsoft.AspNetCore.Mvc.Razor.Extensions.Version2_X/test/ViewComponentTagHelperDescriptorProviderTest.cs
… src/Compiler/Microsoft.AspNetCore.Mvc.Razor.Extensions/test/ViewComponentTagHelperDescriptorProviderTest.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/src/PublicAPI.Unshipped.txt
… src/Compiler/Microsoft.CodeAnalysis.Razor/src/SymbolExtensions.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/src/SymbolVisibilityExtensions.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/src/TagHelperDescriptorProviderContextExtensions.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/src/TagHelperTypes.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/src/WellKnownTypeNames.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/src/WellKnownTypeProvider.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/test/BaseTagHelperDescriptorProviderTest.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/test/BindTagHelperDescriptorProviderTest.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/test/ComponentTagHelperDescriptorProviderTest.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/test/DefaultTagHelperDescriptorFactoryTest.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/test/DefaultTagHelperDescriptorProviderTest.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/test/EventHandlerTagHelperDescriptorProviderTest.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/test/KeyTagHelperDescriptorProviderTest.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/test/RefTagHelperDescriptorProviderTest.cs
… src/Compiler/Microsoft.CodeAnalysis.Razor/test/SplatTagHelperDescriptorProviderTest.cs
… src/Compiler/Microsoft.NET.Sdk.Razor.SourceGenerators/RazorSourceGenerator.TagHelpers.cs
… src/Compiler/Microsoft.NET.Sdk.Razor.SourceGenerators/RazorSourceGenerator.cs
… src/Compiler/Microsoft.NET.Sdk.Razor.SourceGenerators/StaticCompilationTagHelperFeature.cs
… src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/TagHelperResolver.cs
… src/Shared/Microsoft.AspNetCore.Razor.Utilities.Shared/ImmutableArrayExtensions.cs