* Store WellKnownTypeProvider (from dotnet/roslyn-analyzers) in TagHelperDescriptorProviderContext
* Match by type instead of name