@chsienki 2023-02-23
We'll need to run the benchmarks on this to make sure it doesn't regress. Currently the path that we changed name lookup for isn't being hit. So we'll need to incorporate the perf/generator PR on top of this to ensure it doesn't regress in the new scenario.