@chsienki 2023-02-23
@sharwell I merged everything together in order to run the benchmarks, and it looks good.

```
Old:

|           Method |     Mean |   Error |  StdDev |      Gen0 |      Gen1 | Allocated |
|----------------- |---------:|--------:|--------:|----------:|----------:|----------:|
| Cold_Compilation | 391.4 ms | 4.08 ms | 3.82 ms | 6000.0000 | 2000.0000 | 166.92 MB |

|                        Method |     Mean |     Error |    StdDev |    Gen0 |    Gen1 |  Allocated |
|------------------------------ |---------:|----------:|----------:|--------:|--------:|-----------:|
|         Razor_Add_Independent | 5.140 ms | 0.0516 ms | 0.0483 ms | 78.1250 | 31.2500 | 2142.84 KB |
|        Razor_Edit_Independent | 4.143 ms | 0.0417 ms | 0.0370 ms | 85.9375 | 31.2500 |  2186.2 KB |
|      Razor_Remove_Independent | 1.590 ms | 0.0095 ms | 0.0084 ms | 31.2500 | 11.7188 |  818.17 KB |
| Razor_Edit_DependentIgnorable | 2.934 ms | 0.0206 ms | 0.0193 ms | 54.6875 | 15.6250 | 1362.03 KB |
|          Razor_Edit_Dependent | 6.258 ms | 0.1012 ms | 0.0947 ms | 93.7500 | 31.2500 | 2358.26 KB |
|        Razor_Remove_Dependent | 1.542 ms | 0.0308 ms | 0.0329 ms | 31.2500 | 15.6250 |  818.18 KB |


New:

|           Method |     Mean |    Error |   StdDev |   Median |      Gen0 |      Gen1 | Allocated |
|----------------- |---------:|---------:|---------:|---------:|----------:|----------:|----------:|
| Cold_Compilation | 516.2 ms | 10.79 ms | 31.66 ms | 504.8 ms | 5000.0000 | 2000.0000 | 132.29 MB |


|                        Method |     Mean |     Error |    StdDev |   Median |    Gen0 |    Gen1 |  Allocated |
|------------------------------ |---------:|----------:|----------:|---------:|--------:|--------:|-----------:|
|         Razor_Add_Independent | 5.692 ms | 0.1097 ms | 0.0916 ms | 5.718 ms | 46.8750 | 15.6250 | 1405.31 KB |
|        Razor_Edit_Independent | 5.190 ms | 0.1032 ms | 0.3044 ms | 5.109 ms | 54.6875 | 23.4375 | 1434.03 KB |
|      Razor_Remove_Independent | 1.716 ms | 0.0325 ms | 0.0348 ms | 1.713 ms | 15.6250 |  7.8125 |  390.03 KB |
| Razor_Edit_DependentIgnorable | 3.105 ms | 0.0326 ms | 0.0289 ms | 3.096 ms | 39.0625 | 15.6250 | 1023.87 KB |
|          Razor_Edit_Dependent | 7.545 ms | 0.2226 ms | 0.6565 ms | 7.536 ms | 62.5000 | 31.2500 | 1604.64 KB |
|        Razor_Remove_Dependent | 1.724 ms | 0.0329 ms | 0.0308 ms | 1.729 ms | 15.6250 |  3.9063 |  390.04 KB |

```

Note that the 'old' doesn't include https://github.com/dotnet/razor/pull/8297 so the allocations are a little higher, but generally looks like the perf impact is minimal (it was like 10x slower with `GetTypeByMetadataName`)