@333fred 2023-02-24
> I merged everything together in order to run the benchmarks, and it looks good.

@chsienki correct me if I'm reading this wrong, but those benchmarks don't look good? Cold compilation, in particular, looks bad. The minor memory savings don't seem to be paying off with wall-clock time.